+++ SampleStack.Repository/Program.cs
@@ -2,6 +2,8 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using SampleStack.Repository;
+using SampleStack.Repository.Configuration;
+using SampleStack.Repository.Infrastructure;
 
 var host = Host.CreateDefaultBuilder(args)
     .ConfigureAppConfiguration((context, config) =>
